pyproject.toml
[tool.poetry]
name = "diplom-project-api"
version = "0.1.0"
description = ""
authors = ["konstatin fedoseev <kons.fed.99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pytest = "8.4.2"
requests = "^2.32.5"
jsonschema = "^4.25.1"
faker = "^37.12.0"
allure-pytest = "^2.15.2"
selene = "2.0.0rc9"
python-dotenv="1.1.1"
appium-python-client = "^5.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)"
]
